refatcor: Rename to GenerateThreadIdConextIdTestController
